Second-level PEP8 cleanup.
The util.py file has a ton of `undefined name` issues with relation to
what I suspect are supposed to be references to global variables. I did
not clean those up as I don't know how to properly resolve at this time
and I want to keep this specific to PEP8.