pefile swapping Google Code repo for Github.
In 2015 Google retired it's online code repo. As a result pefile moved to Github - https://github.com/erocarrera/pefile/. I have gone ahead an updated the dockerfile to include the new destination of the pefile. In addition I have computed the new SHA256 for the file which has been included. Finally the .tar.gz during decompression creates the directory pefile-pefile-1.2.10-139 which I have also updated on the dockerfile.

I hope this helps!